Use 1px padding on titlebuttons
Summary: This patch fixes titlebutton rendering in Chromium. The headerbar used by Breeze is the same size when a window is maximized, but Chromium uses a smaller headerbar for its tabstrip background when the browser is maximized. When buttons won't fit, Chromium downsizes buttons. This was happening with Breeze titlebuttons since it used 6px of padding on the top and bottom of the buttons, so the buttons appeared taller than they actually were. This patch changes the top and bottom padding to 1px. In native GTK apps, there's no change in appearance.

Reviewers: #breeze, abetts

Reviewed By: abetts

Subscribers: plasma-devel

Tags: #plasma, #breeze

Differential Revision: https://phabricator.kde.org/D9773