Remove hardcoded /usr/share paths in install
Summary:
We can't assume where breeze color files are installed.

If this was a user facing app we should use XDG_DATA_DIRS but as this is
purely used as an install script it should use the data configured in
the found breeze.

There's still a tonne of hardcoded strings in the scripts, but at least
they're now in paths we don't hit.

Test Plan: Now builds on my machine.

Reviewers: #plasma, zzag

Reviewed By: #plasma, zzag

Subscribers: zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18287