Convert colors into GTK colordefs
Summary:
All of the colors pulled from a colorscheme in Breeze GTK now use GTK's color definitons, allowing them to be changed without recompiling the theme.
GTK18 and GTK20 have been merged as well, lessening work required to make changes & maintain.
There are some other minor changes (radio buttons/checkboxes now sink in when clicking as their Qt counterparts do.)
The varaible $new-highlight has been added to gtk.scss. Toggling it to true will cause the highlights to resemble ndavis/highlight @ Breeze. Added that because I had already made the changes to the theme when I did the colordefs, and I don't want to wait on the new highlight style to be set in stone before this can be merged.

One caveat - checkboxes and radiobuttons are hardcoded at build time because they're PNGs. However, the design trend in GNOME applications is to move away from checkboxes and radiobuttons, and they're relatively minor in the grand scheme of things. Redoing them somehow is most likely out of the scope of this revision.

Test Plan: See that the theme works as intended with GTK color definitions.

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D22877