gtk4: remove margin on windows
This isn't /r/unixporn and we want our windows to touch the edges of the screen when maximized. GTK4 is nice enough to automatically adjust the size of surfaces and whatnot when we set a box shadow, so we don't have to use a margin hack in order to make room for it in the first place.

BUG: 431930