assets: resize viewports of titlebutton SVGs
GTK reads SVG to pixmap and creates a cairo surface for an image, and
the size of an image surface used by cairo depends
on the size of SVG viewport but not the size from `background-size`
property. When drawing background images, GTK will scale the pixmap by
width in `background-size` / viewport width, so when the viewport is
smaller than `background-size`, the background image will look blurry.

By making the viewports larger, titlebuttons in GTK
apps (especially in GTK3 Wayland and GTK4) will look much sharper.

CCBUG: 468203