scrollbar: Split margin between scrollbar and trough
The side padding is turned into a margin on the trough to ensure
that GTK WebKit reacts to the full width of the scroll bar.

The vertical padding, however, is kept in the scrollbar to avoid
GTK WebKit scrollbar overflowing past the end of the track.

Likewise for horizontal scroll bars.

Signed-off-by: Sebastian Stark <sebastian.stark@advantest.com>
(cherry picked from commit fad6eb381f1cecb83c7a9c9844bd6082abeb1be3)