scrollbar: Split margin between scrollbar and trough
The side padding is turned into a margin on the trough to ensure
that GTK WebKit reacts to the full width of the scroll bar.

The vertical padding, however, is kept in the scrollbar to avoid
GTK WebKit scrollbar overflowing past the end of the track.

Likewise for horizontal scroll bars.

Signed-off-by: Sebastian Stark <sebastian.stark@advantest.com>